requirements.txt
streamlit==1.28.1
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
pandas==2.1.3
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
plotly==5.17.0
scikit-learn==1.3.2
tiktoken==0.5.1
pydantic==2.5.0
